fix(api): moved and renamed func
refs. metriport/metriport-internal#2240
